feat: Refactor import statements and remove unused code #1154
This commit includes a refactor in the import statements of fitbitter.dart,
where we now group them by type (Data models, Errors, Managers, URLs, Connector).
We added fitbitActivityTimeseriesIntradayDataManager and
fitbitActivityTimeseriesIntradayAPIURL to their respective categories. We also
removed fitbitAccountAPIURL as it was not being used.

In fitbitActivityTimeseriesIntradayAPIURL.dart, we removed the unused enum
Resource and its corresponding map resourceToString, and refactored the import
statements.